Optimize and clean-up (#48)
